requirements.txt
# Core video processing
scenedetect[opencv]>=0.6.2
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Machine Learning & Embeddings - Pinned to compatible versions
torch==2.0.1+cu118
torchvision==0.15.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

# CLIP (install after torch)
git+https://github.com/openai/CLIP.git

# Vector Storage & Search
faiss-cpu>=1.7.4
numpy==1.24.4
pandas>=1.5.0

# Utilities
pillow>=9.0.0
python-dotenv>=1.0.0

# OpenAI for GPT-4 embeddings and enhanced text understanding
openai>=1.0.0
